Updated WebRTC_vad
The Signal to Noise Ratio voice activity detector (SNR_VAD) is
working much better than I expected it to, and WebRTC_VAD is more
sensitive than I expected it to be.

In my continuing quest to eliminate false-positives in wake word
detection, I decided to update the WebRTC_VAD plugin to add the
WebRTC stuff on top of the SNR stuff, so the WebRTC_VAD plugin
is really using both methods.

I also increased the default "aggressiveness" level of WebRTC_VAD
from 1 to the maximum of 3.

WebRTC_VAD is still finicky about chunk size so I suggest
continuing to use the simpler SNR_VAD as the default.