Webrtc vad update (#338)
* Various fixes

This includes a fix for multiple keywords and some small changes to
the NaomiSTTTrainer.py app.

* Removed extra parameter

Removed an extra "true" from a call to simple input for
"allow multiple", a setting I just removed and will deal with at
the setting level instead.

* Settings for notificationclient

Added settings for the demo notification client so it now will
prompt the user for a list of email addresses it is allowed to
respond to.

Also changed the "allow_multiple" parameter for settings to
"return_type", which I think will be more flexible. In this case,
you can set the return type to "list" which will convert your
string response into a list of strings. It can also return values
as boolean, number, or str (the default). This currently only
works with the default input control type, although it could
easily be extended to other types.

* Fix "noise" output

* Remove unused variable

I'm trying to figure out what is the more flexible option - a
return_type parameter, or an accept_list parameter. The
return_type parameter seems more flexible, except that the type
of response is implied in the name of the control (text, boolean,
number, etc). If I set it so that a boolean control actually
returns a boolean or a number returns a number, then I can use
"accept_list" to return a list of booleans or a list of numbers.

* Change back to "return_list"

* Fix greeting

When using multiple keywords, Naomi greets the user with "My name
is <name1> but you can also call me <name2> or <name3>". This
fixed the "or" join so "or" was not appended to the previous name.

* Updated WebRTC_vad

The Signal to Noise Ratio voice activity detector (SNR_VAD) is
working much better than I expected it to, and WebRTC_VAD is more
sensitive than I expected it to be.

In my continuing quest to eliminate false-positives in wake word
detection, I decided to update the WebRTC_VAD plugin to add the
WebRTC stuff on top of the SNR stuff, so the WebRTC_VAD plugin
is really using both methods.

I also increased the default "aggressiveness" level of WebRTC_VAD
from 1 to the maximum of 3.

WebRTC_VAD is still finicky about chunk size so I suggest
continuing to use the simpler SNR_VAD as the default.