Verify Wakeword
This allows you to tell Naomi to use the active stt engine to
verify that it heard the wakeword before responding. This greatly
reduces the number of false positives (Naomi responding when not
addressed) but may cause Naomi to miss times when you do say the
wake word (false rejection).

It adds ["passive_stt", "verify_wakeword"] and
["listen_while_speaking"] options to the initial setup script.

This update also changes the MPD Control plugin to respond
directly to requests instead of going into a special music
"mode" where normal requests are rejected, so now you can ask
what time it is while listening to a playlist.

It also allows the stop speechhandler to search for any
speechhandler with a "stop" method and run all of them when you
ask Naomi to stop. This way you can stop any plugin that is
generating noise with the single command: "Naomi, stop!"