Initial configuration fixes
This contains a few bug fixes related to my experience installing
Naomi to a brand-new Raspberry Pi 4B running Raspberry Pi OS
(2021-05-07-raspios-buster-armhf-lite.img)

These include minor annoyances like setting the _input_rate,
_input_bits, _input_channels and _input_chunksize properties of
the selected input device to int when reading from the profile.

It goes on to include some additional settings, like audio:input_rate
and audio:input_chunksize. The later is import if you select
webrtc_vad as your Voice Activity Detector since it requires
that audio chunks be audio:input_rate bytes/second * .03 seconds.

I changed the default value for "passive_stt:verify_wakeword" from
True to False, since this setting makes the default configuration
using Pocketsphinx for both passive and active speech to text
reject almost every utterance.

I changed the way settings that are skipped (active=False) are
written to profile.yml. Instead of simply putting an empty string
as a placeholder, the current default value is put. This probably
needs to be updated so that the plugin __init__ method ignores
inactive settings when deciding if the plugin needs to be
configured, but I'm not sure what the downsides of that will be.

I also made it so the pluginstore is written to a static variable
in the naomi/profile class, so we don't have to keep searching for
plugins every time we want to search for a class of plugins or
initialize an instance of a plugin.

I went ahead and stored all settings to a naomi.profile structure
called _settings. This both allows us to keep track of which
settings have already been addressed so we don't have to configure
a plugin like pocketsphinx that may be initialized multiple times
the same number of times it is initialized, and will allow us to
quickly look up the settings configuration when settings are
configured non-sequentially (such as with an html interface).