Configuration fixes (#344)
* Various fixes

This includes a fix for multiple keywords and some small changes to
the NaomiSTTTrainer.py app.

* Removed extra parameter

Removed an extra "true" from a call to simple input for
"allow multiple", a setting I just removed and will deal with at
the setting level instead.

* Settings for notificationclient

Added settings for the demo notification client so it now will
prompt the user for a list of email addresses it is allowed to
respond to.

Also changed the "allow_multiple" parameter for settings to
"return_type", which I think will be more flexible. In this case,
you can set the return type to "list" which will convert your
string response into a list of strings. It can also return values
as boolean, number, or str (the default). This currently only
works with the default input control type, although it could
easily be extended to other types.

* Fix "noise" output

* Remove unused variable

I'm trying to figure out what is the more flexible option - a
return_type parameter, or an accept_list parameter. The
return_type parameter seems more flexible, except that the type
of response is implied in the name of the control (text, boolean,
number, etc). If I set it so that a boolean control actually
returns a boolean or a number returns a number, then I can use
"accept_list" to return a list of booleans or a list of numbers.

* Change back to "return_list"

* Fix greeting

When using multiple keywords, Naomi greets the user with "My name
is <name1> but you can also call me <name2> or <name3>". This
fixed the "or" join so "or" was not appended to the previous name.

* Updated WebRTC_vad

The Signal to Noise Ratio voice activity detector (SNR_VAD) is
working much better than I expected it to, and WebRTC_VAD is more
sensitive than I expected it to be.

In my continuing quest to eliminate false-positives in wake word
detection, I decided to update the WebRTC_VAD plugin to add the
WebRTC stuff on top of the SNR stuff, so the WebRTC_VAD plugin
is really using both methods.

I also increased the default "aggressiveness" level of WebRTC_VAD
from 1 to the maximum of 3.

WebRTC_VAD is still finicky about chunk size so I suggest
continuing to use the simpler SNR_VAD as the default.

* In passive mode, verify keyword

With this change, if Naomi hears any wake word using the passive
stt engine, it will verify that the active transcription also
contains the same wake word. This cuts down a lot on false
positives.

* Verify Wakeword

This allows you to tell Naomi to use the active stt engine to
verify that it heard the wakeword before responding. This greatly
reduces the number of false positives (Naomi responding when not
addressed) but may cause Naomi to miss times when you do say the
wake word (false rejection).

It adds ["passive_stt", "verify_wakeword"] and
["listen_while_speaking"] options to the initial setup script.

This update also changes the MPD Control plugin to respond
directly to requests instead of going into a special music
"mode" where normal requests are rejected, so now you can ask
what time it is while listening to a playlist.

It also allows the stop speechhandler to search for any
speechhandler with a "stop" method and run all of them when you
ask Naomi to stop. This way you can stop any plugin that is
generating noise with the single command: "Naomi, stop!"

* Fixed profile numbers

Two problems with number fields - first, a changed value was not
getting recorded, second, a changed value was being recorded as
a string, so it needs to always be converted to an int or float
as needed when retrieved from the profile.

* Initial configuration fixes

This contains a few bug fixes related to my experience installing
Naomi to a brand-new Raspberry Pi 4B running Raspberry Pi OS
(2021-05-07-raspios-buster-armhf-lite.img)

These include minor annoyances like setting the _input_rate,
_input_bits, _input_channels and _input_chunksize properties of
the selected input device to int when reading from the profile.

It goes on to include some additional settings, like audio:input_rate
and audio:input_chunksize. The later is import if you select
webrtc_vad as your Voice Activity Detector since it requires
that audio chunks be audio:input_rate bytes/second * .03 seconds.

I changed the default value for "passive_stt:verify_wakeword" from
True to False, since this setting makes the default configuration
using Pocketsphinx for both passive and active speech to text
reject almost every utterance.

I changed the way settings that are skipped (active=False) are
written to profile.yml. Instead of simply putting an empty string
as a placeholder, the current default value is put. This probably
needs to be updated so that the plugin __init__ method ignores
inactive settings when deciding if the plugin needs to be
configured, but I'm not sure what the downsides of that will be.

I also made it so the pluginstore is written to a static variable
in the naomi/profile class, so we don't have to keep searching for
plugins every time we want to search for a class of plugins or
initialize an instance of a plugin.

I went ahead and stored all settings to a naomi.profile structure
called _settings. This both allows us to keep track of which
settings have already been addressed so we don't have to configure
a plugin like pocketsphinx that may be initialized multiple times
the same number of times it is initialized, and will allow us to
quickly look up the settings configuration when settings are
configured non-sequentially (such as with an html interface).

* Fix default value

I changed this to set the profile variable to the default value
if the active property is False, but the default value was only
being evaluated after the "active" property was, so this was
generating a "variable 'default' referenced before being defined'
error message