Better error messages (#349)
This improves the error message handling when importing plugins.
Currently, they almost all list "unknown" as the reason they are
being skipped, when in fact it is known. This is because any that
are skipped because of missing modules are getting a custom
exception which does not have a message property. It turns out
that these display nicely if you just use str() to convert them
into strings. Since other error types also provide useful
information this way, I am degrading to str(e) instead of just
"Unknown" if e does not have message or msg properties.

I also added "please" to the list of standard phrases, since I use
it quite often and Naomi tries to interpret it as other in-grammar
words which never works out well.

Finally, wwis_weather will provide a more useful error message if
you have not configured a location, rather than just suggesting
you might want to look out a window.