Added a default speaker recognizer
Added a default speaker recognizer, default_sr, which does not
attempt to identify the speaker but just responds with the first
name as stored in the profile.

I am also passing the result around from the sr_plugin to the
intent parser, to which the identity of the speaker can be
attached to the intent object being passed to the speechhandler.

I also simplified the list of parameters being passed to the mic
object when it is created, storing them instead in the profile.