Fixed problems with speech recognizer
There were numerous problems with the speech recognizer class that
made it not work when using the VOSK_sr plugin.

I am no longer trying to recognize the speaker when listening
passively for the wake word, only when doing the active listening.
This is because the passive listening needs to be very fast.

I am now putting the name of the identified user in parenthesis
after the active listening transcript. Plugins can access the
identity of the speaker as `intent.get('user', '')`. The only
plugin currently set up to use this is the shutdown plugin. I
also have an update to the Greetings plugin which greets you by
name when you greet Naomi.

The setup still assumes en-US when downloading the VOSK models,
which needs to be fixed to respect the "language" setting in the
profile.

The VOSK speaker recognition is not terribly accurate. It also
seems like you need to retrain your speaker recognition database
from new recordings when you switch to different recording
hardware.