Download French or German language models
Added filenames for the default French and German models.
Constructed URLs and paths from these models to automatically
download and extract the model that matches the language choice
in profile.

The only thing it does not currently do is check whether or not
VOSK is also used as the STT engine. If not, then the audio
file pointer should be passed to the actual STT engine.