Reinit Pocketsphinx vocabularies
I am working on a Libre ROC-RK3328-CC Renegade computer as my main
development board for Naomi. I am suddenly getting this error
after using "expect" (in the joke plugin):

ERROR:naomi.mic:Passive transcription failed!
Traceback (most recent call last):
  File "/home/naomi/Naomi/naomi/mic.py", line 318, in wait_for_keyword
    transcribed = [word.upper() for word in self.passive_stt_engine.transcribe(f)]
  File "/home/naomi/Naomi/plugins/stt/pocketsphinx-stt/sphinxplugin.py", line 417, in transcribe
    self._decoder.process_raw(data, False, True)
  File "_pocketsphinx.pyx", line 960, in _pocketsphinx.Decoder.process_raw
IndexError: Out of bounds on buffer access (axis 0)

I'm not sure what this error means. It always seems to affect the
passive listening engine. Usually I use the pocketsphinx_kws plugin
for passive listening, so this may have been happening for a while
and I just didn't notice. Re-initializing the engine seems to work
fine, so I'm just keeping an eye on it. The biggest problem with
re-initializing is that with our current audio system, Naomi tends
to stutter when the re-initialization happens. This is one reason
for wanting to move to a multi-threaded approach and see if that
helps.