Fix "Too many open files" error
Recently, I am getting a "Too many open files" error. This appears
to be because a file handle is opened to /dev/null so STDERR can
be redirected to it before performing some operations in
Pocketsphinx, but never getting closed. This gives the io.open
method context so it will close automatically when it goes out
of scope.

Also fixed an issue reported by CodeQL where a user-provided path
is being used to serve content.