This environment is externally managed
I just tried installing Naomi on my new Bookworm computer and got
the above error message. It seems that Python now allows the OS
to lock down the main python environment so you can't install
to it using pip or conda or whatnot. You have to use the OS
repositories if you want to install packages on the main python
environment.

This is good, and I'm sure it's the right way to do it. But
although almost all of our packages are installed within the
Naomi virtual environment, we have been using pip to install
virtualenvwrapper itself to the main python environment as a
user package. Since we can't, the install script breaks.

To fix this issue, I had to add the python3-virtualenvwrapper
package to the apt_requirements.txt file, then remove the pip3
install --user virtualenvwrapper line from script.deb.sh.

The location of the virtualenvwrapper.sh script changes from
~/.local/bin/virtualenvwrapper.sh to
/usr/share/virtualenvwrapper/virtualenvwrapper.sh
since it is now being installed by the OS.

The location of virtualenv also changes. In this case, it
worked to simply remove all the VIRTUALENVWRAPPER_VIRTUALENV
lines.

I also made a small change to sphinxplugin.py where I am no
longer checking for the g2p fst model when deciding whether to
clone the CMUSphinx standard language model, since that file
is generated after the model is downloaded. If something
happens between downloading the language model and generating the
g2p model things would get confused.