Add progress bar to download_file()
Add a progress bar to app_utils.download_file(url, local_filename)
function. This is only really useful when using the command line
client (we don't have another client right now) and only before
the microphone feedback starts to display. It would be nice to
move this into a visualization someday, but this is pretty tightly
integrated with the operation itself.

This commit also gives the download_file function the ability to
resume partial downloads.