Remove default location from logfile (#399)
In my last request, I accidentally left the logfile with a default,
so Naomi still records log entries to that file. Also, my idea for
having two log handlers - one that writes to a file for anything
below ERROR level messages and another that also sends ERROR and
CRITICAL messages to the screen did not work because the handler
settings do not automatically come with the logger, just the
settings passed to logging.basicConfig do. So every time a logger
is created, the handlers have to be added.