Synchronous and Asynchronous Mic (#402)
* Synchronous and Asynchronous Mic

I have turned the Mic class into an abstract class and used it to
create two new classes, MicSynchronous and MicAsynchronous. I'm
hoping to expand it to all the different mic classes, including
the local (text) mic and the batch mic.

I'm attempting to support both active listen mode (where the
computer only starts listening for a command after hearing its
wakeword - Siri-like mode) and passive listen mode (where the
computer records blocks of audio, then checks for the wake word
and then checks the same block of audio for a command - Echo-like).

Right now, I am having trouble with the expect function when using
passive listen mode with the asynchronous listener. This has to do
with the pyaudio device play_file, which returns when it has
finished writing to the queue, but before the audio is done
playing. This leads to situations where the next audio starts
getting queued before the last audio finishes playing. If the
audio's have different frame sizes, this leads to a segmentation
fault.

I have been testing by using the "knock knock joke" and "time"
speechhandler plugins. Knock-knock joke uses expect quite a bit.
I have been using Pocketsphinx_KWS for my passive STT engine,
Pocketsphinx for my special STT engine and VOSK (which is
available here: https://github.com/aaronchantrill/Naomi_VOSK_STT)
as my active STT engine. VOSK works well, at least in English,
but requires some additional training if you have non-standard
words in your vocabulary. I'd like to make VOSK officially
available through NPE but the last time I trained VOSK to
recognize some additional words, it required a computer with
32GiB of ram. I will test on my Raspberry Pi 5 with 8 GiB and
see if it can handle it, but have low expectations. I would
like to add an option to export the Naomi vocabulary so VOSK
can be trained on another computer, as it does run well on the
Raspberry Pi 4 and 5.

* Small fixes for flake8

* Fix Mic initialization

Mic class is a child class of i18n.GettextMixin but was not using
that parent class correctly. Edited the __init__ method.

* Added Utterance class

I added an Utterance class that returns information from the mic
listen() and active_listen() methods. This object returns the
transcription when called without an parameters, so it is
backwards compatible with plugins that expect to get a string or
list of strings back from those methods.