These are some small fixes
Probably the most notable here is the breaking change of changing
the "verify_wakeword" argument to "verify_keyword". This really
isn't an important change, and it may cause some issues with the
VOSK plugin depending on which version is installed.

This fixes an issue where the asyncronous mic assumes you are
using verify_keywords even if you are not.

Finally, this sets the variable keyword to a string and keywords
to a list in application.py, and saves the keywords to an arg
variable.