Fix issue with empty profile.yml file
When using yaml.safe_load() on an empty text file, the returned
value is None rather than an empty dictionary as I expected. This
caused an error when adding the first value, as I would get the
error message "NoneType object does not support item assignment".

This patch changes a None value into an empty dictionary.