Add an array setting (#417)
This is a special type of setting for a set of related values that
can be repeated an arbitrary number of times. It requires an "each"
property, which contains the list of subsettings to be repeated.
An example would be when adding a number of custom lights from
HomeAssistant. Each light needs an ID which allows Naomi to tell
HomeAssistant which light is being controlled, plus a friendly
name that lets the user request that particular light:

```
("SimpleHomeAssistant", "light"), {
        "type": "array",
        "each": OrderedDict([
            (
                ("id",), {
                    "title": _("ID of your light"),
                    "description": _("The ID of your light from HomeAssistant")
                }
            ),
            (
                ("name",), {
                    "title": _("Name of your light"),
                    "description": _("The name you would like to use to refer to your light.")
                }
            )
        ])
    }
)
```

will allow you to create a structure like:

```
  light:
  - id: light1
    name: cottage light
  - id: light2
    name: test light
```
in your profile.yml file.