Play greeting before starting mic (#418)
I realized that one of the big problems with Naomi listening to
itself speak is that it says all of its keywords/wakewords as
part of the first message, which causes it to scan its greeting
for intent and causes it to respond with a message immediately,
which isn't great.

I added a say_sync method to the mic.py interface and used that
directly in conversation.py. I also changed the askName method
to sayName since that's what it actually does now.

I also made a small change to plugin.py so that the messages
about which plugin it is configuring when configuring plugins
appears highlighted.

Finally, I made a small change to commandline.py so that if a
listbox setting has only one option, that option will be selected
by default.

Small changes, nothing earth shattering here.