Fix adding new settings array
This fixes an issue with plugins that use the new "array" setting
type. The array setting type allows you to specify an arbitrary
number of items for a setting. For instance, you could define
"user" as type "array" and then have a name and email address
defined for each user.

The issue is that when first creating the setting, I was treating
the subsettings like dictionaries, but they are actually a tuple
of tuples, with the first representing the path and the second
the definition.

The new "array" type setting is necessary, but overly complex. I
wish I could come up with a simpler implementation.

Anyway, this fixes a "TypeError" error when attempting to prompt
the user for a value for an "array" setting that has not been
initialized before.