Allow Naomi to use an LLM
This change basically gives Naomi the ability to use an LLM in
crafting responses. The LLM is contacted over a web api, so any
OpenAI compatible LLM server should be fine. I am currently using
llama.cpp on a system with an nVidia GeForce RTX 3060 with 12 GB
VRAM. I am currently using the following model:
https://huggingface.co/mav23/Llama_3.2_1B_Intruct_Tool_Calling_V2-GGUF/blob/main/llama_3.2_1b_intruct_tool_calling_v2.Q4_K_M.gguf
This is only a 3B parameter model, so it should run fine on cards
with less VRAM. I have also tried running it on an Intel graphics
card on my laptop using SyCL which seems to work fine.

This is very much experimental. The basic idea is that I use the
existing TTI parser to activate a plugin, then pass the output
from that plugin to the LLM as a system message. This allows the
LLM to use current data and answer questions like "What time is
it?" or "Are we expecting rain on Thursday?" I'm also looking into
using function calling which would allow the LLM to be used as
an intent parser and might make plugin activation more integrated,
but that has its own set of problems, including that different
models define and use plugins differently.

I also have included some notebooks. I plan to use these to create
a set of notebooks that will hopefully be helpful to people who
want to better understand how Naomi works internally.